Disable keyword style, so that identifiers like :: are accepted
In older CHICKENs, keywords could be bound like regular symbols, which
was a bit of a strange behaviour.  In 5.0.2 and later this no longer
works, so we'll need to either quote the symbols using pipe characters
or disable the keyword style reader.

Given that this code doesn't seem to use keywords at all, simply
disabling the reader for keyword is the simplest fix.